requirements.txt
ttkbootstrap
spotipy
Pillow
requests
pyinstaller
tinyaes 